fix Modal
